add product page, guest functions

+++ web-application/src/features/homepage/context/CartContext.jsx
@@ -3,6 +3,7 @@ import { createContext, useState } from "react";
 import CartReducer from "./cartReducer";
 import cartApi from "../../../apis/cart";
 import { getAccessToken } from "../../../utils/local-storage";
+import useAuth from "../../../hooks/useAuth";
 
 export const CartContext = createContext();
 
@@ -15,8 +16,9 @@ export const CartContext = createContext();
 
 export default function CartContextProvider({ children }) {
   const [cart, dispatch] = useReducer(CartReducer, []);
-
   const [authorized, setAuthorized] = useState(false);
+  const { authUser } = useAuth();
+  const [refresh, setRefresh] = useState(false); // ให้ re-render ใหม่เมื่อมีการอัพเดทสินค้าในtoken
 
   const getAllProductsinCart = async () => {
     try {
@@ -41,20 +43,51 @@ export default function CartContextProvider({ children }) {
   }, []);
 
   const handleQuantity = async (data) => {
-    try {
-      await cartApi.adjustItemNumber(data);
-      getAllProductsinCart();
-    } catch (error) {
-      console.log(error);
+    if (authUser) {
+      // ถ้าผู้ใช้ล็อกอินอยู่, เรียกใช้ API
+      try {
+        await cartApi.adjustItemNumber(data);
+        getAllProductsinCart();
+      } catch (error) {
+        console.log(error);
+      }
+    } else {
+      // ถ้าผู้ใช้เป็น guest, ปรับข้อมูลใน local storage
+      const guestCart = JSON.parse(localStorage.getItem("guestCart")) || [];
+      const itemIndex = guestCart.findIndex(
+        (item) => item.productId === data.productId
+      );
+
+      if (itemIndex > -1) {
+        guestCart[itemIndex].amount += data.amount;
+        if (guestCart[itemIndex].amount <= 0) {
+          guestCart.splice(itemIndex, 1); // ลบสินค้าถ้าจำนวน <= 0
+        }
+      }
+
+      localStorage.setItem("guestCart", JSON.stringify(guestCart));
+      setRefresh(!refresh);
     }
   };
 
   const handleRemoveItem = async (cartItemId) => {
-    try {
-      await cartApi.deleteCartItem(cartItemId);
-      getAllProductsinCart();
-    } catch (error) {
-      console.log(error);
+    if (authUser) {
+      // ถ้าผู้ใช้ล็อกอินอยู่, เรียกใช้ API
+      try {
+        await cartApi.deleteCartItem(cartItemId);
+        getAllProductsinCart();
+      } catch (error) {
+        console.log(error);
+      }
+    } else {
+      // ถ้าผู้ใช้เป็น guest, ลบสินค้าใน local storage
+      const guestCart = JSON.parse(localStorage.getItem("guestCart")) || [];
+      const updatedCart = guestCart.filter(
+        (item) => item.productId !== cartItemId
+      );
+
+      localStorage.setItem("guestCart", JSON.stringify(updatedCart));
+      setRefresh(!refresh);
     }
   };
